Move from events to activities.
According to the documentation, "events" is
reserved for talks organised by the CCCS,
while "activites" is for talks with invited
speakers from the CCCS.

This talk does not fit in any of these
categories, but it fits better in the latter.
Former "interesting" talks have been
"activities", too.

Please review.

Signed-off-by: Thomas Hochstein <thh@inter.net>